requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
sqlparse==0.4.3
numpy==1.24.2
django-colorfield==0.8.0
Pillow==9.4.0
djangorestframework==3.14.0
